Bookings: use dataloader to fetch all bookings
We used some kind of weird memoization but it didn't work anyway. This fix will send only one request when sending queries like this one:

```
query ManageMyBooking {
  future: customerBookings(only: FUTURE) {
    edges {
      node {
        destinationImageUrl
      }
    }
  }
  past: customerBookings(only: PAST) {
    edges {
      node {
        destinationImageUrl
      }
    }
  }
}
```

We are currently sending 2 requests for this and backend for some reason returns error code for it. This issue should be therefore fixed as well (hopefully).